+++ src/hooks/useSignupMutation.tsx
@@ -1,33 +1,9 @@
 import { useSetRecoilState } from 'recoil';
-import HttpClient from '../apis/HttpClient';
-import { IAuthProps } from '../service/auth/type';
 import { useMutation } from '@tanstack/react-query';
 import { loginSuccessState } from '../recoil/atom/loginSuccessState';
 import { useProfileState } from '../context/UserProfile';
 import { setAccessTokenStorage, setRefreshTokenStorage } from '../repository/AuthRepository';
-
-export async function SignUpService({
-    email,
-    categoryIds,
-    nickName,
-    provider,
-    providerId,
-    profileImageUrl,
-}: IAuthProps) {
-    try {
-        const res = await HttpClient.post(`/api/v1/auth/signup`, {
-            email: `${email}`,
-            categoryIds: categoryIds.map(Number),
-            nickname: `${nickName}`,
-            provider: `${provider}`,
-            providerId: `${providerId}`,
-            profileImageUrl: `${profileImageUrl}`,
-        });
-        return res.data;
-    } catch (error) {
-        console.error(error);
-    }
-}
+import { SignUpService } from '../service/auth/SocialService';
 
 export const useSignUpMutation = () => {
     const setLoginSucess = useSetRecoilState(loginSuccessState);
