+++ src/hooks/useTopPostQuery.tsx
@@ -6,7 +6,6 @@ export async function getTopPostItemService() {
     const TYPE = 'VIEWS';
     try {
         const res = await HttpClient.get(`api/v1/posts/top/${TYPE}/${COUNT}`);
-        console.log(res);
         return res.data;
     } catch (error) {
         console.error(error);
